add retries on bridge and master connection, small refactoring
